Added sdimg-ubi selection to imagetypes
With a patch to assemble-image.sh[1], this change to the user interface
enables a copy of the root file system to be built into a UBIFS image
and stored on the SD card for use in flashing systems with the same root
file system.

It is possible to make this a checkbox that would show up when sdimg is
selected, but that would complicate the logic quite a bit.  There seems
to be a reasonable amount of space on the line, so the simpler logic
seemed like the better trade-off.

[1] http://lists.linuxtogo.org/pipermail/angstrom-distro-devel/2011-July/004778.html

Signed-off-by: Koen Kooi <koen@dominion.thruhere.net>